+++ src/spi.rs
@@ -36,10 +36,9 @@ pub struct Mode {
 mod hal_02;
 mod hal_1;
 
-use crate::pac::spi1;
-use crate::rcc;
+use crate::pac::{spi1, RCC};
+use crate::rcc::{self, Rcc};
 
-use crate::rcc::Clocks;
 use enumflags2::BitFlags;
 use fugit::HertzU32 as Hertz;
 
@@ -256,7 +255,7 @@ pub trait SpiExt: Sized + Instance {
         ),
         mode: impl Into<Mode>,
         freq: Hertz,
-        clocks: &Clocks,
+        rcc: &mut Rcc,
     ) -> Spi<Self, false, u8>;
 
     fn spi_bidi(
@@ -264,7 +263,7 @@ pub trait SpiExt: Sized + Instance {
         pins: (impl Into<Self::Sck>, impl Into<Self::Mosi>),
         mode: impl Into<Mode>,
         freq: Hertz,
-        clocks: &Clocks,
+        rcc: &mut Rcc,
     ) -> Spi<Self, true, u8>
     where
         NoPin: Into<Self::Miso>;
@@ -278,6 +277,7 @@ pub trait SpiExt: Sized + Instance {
             Option<Self::Nss>,
         ),
         mode: impl Into<Mode>,
+        rcc: &mut RCC,
     ) -> SpiSlave<Self, false, u8>;
 
     fn spi_bidi_slave(
@@ -288,6 +288,7 @@ pub trait SpiExt: Sized + Instance {
             Option<Self::Nss>,
         ),
         mode: impl Into<Mode>,
+        rcc: &mut RCC,
     ) -> SpiSlave<Self, true, u8>
     where
         NoPin: Into<Self::Mosi>;
@@ -308,9 +309,9 @@ impl<SPI: Instance> SpiExt for SPI {
         ),
         mode: impl Into<Mode>,
         freq: Hertz,
-        clocks: &Clocks,
+        rcc: &mut Rcc,
     ) -> Spi<Self, false, u8> {
-        Spi::new(self, pins, mode, freq, clocks)
+        Spi::new(self, pins, mode, freq, rcc)
     }
     /// Enables the SPI clock, resets the peripheral, sets `Alternate` mode for `pins` and initialize the peripheral as SPI Master BIDI mode.
     ///
@@ -322,12 +323,12 @@ impl<SPI: Instance> SpiExt for SPI {
         pins: (impl Into<Self::Sck>, impl Into<Self::Mosi>),
         mode: impl Into<Mode>,
         freq: Hertz,
-        clocks: &Clocks,
+        rcc: &mut Rcc,
     ) -> Spi<Self, true, u8>
     where
         NoPin: Into<Self::Miso>,
     {
-        Spi::new_bidi(self, pins, mode, freq, clocks)
+        Spi::new_bidi(self, pins, mode, freq, rcc)
     }
     /// Enables the SPI clock, resets the peripheral, sets `Alternate` mode for `pins` and initialize the peripheral as SPI Slave Normal mode.
     ///
@@ -343,8 +344,9 @@ impl<SPI: Instance> SpiExt for SPI {
             Option<Self::Nss>,
         ),
         mode: impl Into<Mode>,
+        rcc: &mut RCC,
     ) -> SpiSlave<Self, false, u8> {
-        SpiSlave::new(self, pins, mode)
+        SpiSlave::new(self, pins, mode, rcc)
     }
     /// Enables the SPI clock, resets the peripheral, sets `Alternate` mode for `pins` and initialize the peripheral as SPI Slave BIDI mode.
     ///
@@ -359,11 +361,12 @@ impl<SPI: Instance> SpiExt for SPI {
             Option<Self::Nss>,
         ),
         mode: impl Into<Mode>,
+        rcc: &mut RCC,
     ) -> SpiSlave<Self, true, u8>
     where
         NoPin: Into<Self::Mosi>,
     {
-        SpiSlave::new_bidi(self, pins, mode)
+        SpiSlave::new_bidi(self, pins, mode, rcc)
     }
 }
 
@@ -478,17 +481,15 @@ impl<SPI: Instance> Spi<SPI, false, u8> {
         ),
         mode: impl Into<Mode>,
         freq: Hertz,
-        clocks: &Clocks,
+        rcc: &mut Rcc,
     ) -> Self {
-        unsafe {
-            SPI::enable_unchecked();
-            SPI::reset_unchecked();
-        }
+        SPI::enable(rcc);
+        SPI::reset(rcc);
 
         let pins = (pins.0.into(), pins.1.into(), pins.2.into());
 
         Self::_new(spi, pins)
-            .pre_init(mode.into(), freq, SPI::clock(clocks))
+            .pre_init(mode.into(), freq, SPI::clock(&rcc.clocks))
             .init()
     }
 }
@@ -504,20 +505,18 @@ impl<SPI: Instance> Spi<SPI, true, u8> {
         pins: (impl Into<SPI::Sck>, impl Into<SPI::Mosi>),
         mode: impl Into<Mode>,
         freq: Hertz,
-        clocks: &Clocks,
+        rcc: &mut Rcc,
     ) -> Self
     where
         NoPin: Into<SPI::Miso>,
     {
-        unsafe {
-            SPI::enable_unchecked();
-            SPI::reset_unchecked();
-        }
+        SPI::enable(rcc);
+        SPI::reset(rcc);
 
         let pins = (pins.0.into(), NoPin::new().into(), pins.1.into());
 
         Self::_new(spi, pins)
-            .pre_init(mode.into(), freq, SPI::clock(clocks))
+            .pre_init(mode.into(), freq, SPI::clock(&rcc.clocks))
             .init()
     }
 }
@@ -537,11 +536,10 @@ impl<SPI: Instance> SpiSlave<SPI, false, u8> {
             Option<SPI::Nss>,
         ),
         mode: impl Into<Mode>,
+        rcc: &mut RCC,
     ) -> Self {
-        unsafe {
-            SPI::enable_unchecked();
-            SPI::reset_unchecked();
-        }
+        SPI::enable(rcc);
+        SPI::reset(rcc);
 
         let pins = (pins.0.into(), pins.1.into(), pins.2.into(), pins.3);
 
@@ -559,14 +557,13 @@ impl<SPI: Instance> SpiSlave<SPI, true, u8> {
         spi: SPI,
         pins: (impl Into<SPI::Sck>, impl Into<SPI::Miso>, Option<SPI::Nss>),
         mode: impl Into<Mode>,
+        rcc: &mut RCC,
     ) -> Self
     where
         NoPin: Into<SPI::Mosi>,
     {
-        unsafe {
-            SPI::enable_unchecked();
-            SPI::reset_unchecked();
-        }
+        SPI::enable(rcc);
+        SPI::reset(rcc);
 
         let pins = (pins.0.into(), pins.1.into(), NoPin::new().into(), pins.2);